Updates the <see cref="ProductPart.CanBeBought"/> dictionary and applies the changes to the SKU key of <see cref="ProductAttributeField"/>
