+++ src/Modules/OrchardCore.Commerce/ViewModels/ProductPartViewModel.cs
@@ -23,5 +23,7 @@ public class ProductPartViewModel : ISkuHolderContent
     public IEnumerable<ProductAttributeDescription> Attributes { get; set; }
 
     [BindNever]
-    public IDictionary<string, bool> CanBeBought { get; private set; } = new Dictionary<string, bool>();
+#pragma warning disable CA2227
+    public IDictionary<string, bool> CanBeBought { get; set; } = new Dictionary<string, bool>();
+#pragma warning restore CA2227
 }
